Added OnClickListener to CardView to Navigate to Details

+++ app/src/main/java/com/example/blockbuster/network/Movie.kt
@@ -11,5 +11,5 @@ data class Movie(
     @Json(name = "release_date") val releaseDate: String,
     @Json(name = "poster_path") val imgSrcUrl: String,
     @Json(name = "backdrop_path") val backdrop: String,
-    val popularity: Double
+    val popularity: Float
 )
